implemeted the test for the app
