+++ package.json
@@ -14,6 +14,7 @@
     "axios": "^0.19.0",
     "express": "^4.17.1",
     "jest": "^25.1.0",
-    "nodemon": "^2.0.2"
+    "nodemon": "^2.0.2",
+    "supertest": "^4.0.2"
   }
 }
