--- test/inedex.test.js
@@ -1,3 +1,0 @@
-test("The test",()=>{
-    expect(1).toBe(1);
-})
